backend/pyproject.toml
[project]
name = "job-search-planner"
version = "0.1.0"
description = "Daily planner for job search activities"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
]
